Moved implementation to lib functions, made various improvements, added thiserror dependency for future use
